+++ src/tests/files_test.py
@@ -2,7 +2,7 @@ import ast
 import os
 
 from tryceratops.files.parser import parse_file
-from tryceratops.filters import IgnoreViolation
+from tryceratops.settings import IgnoreViolation
 
 
 def get_full_path(filename: str):
