+++ src/tryceratops/__main__.py
@@ -4,10 +4,10 @@ from typing import Tuple
 
 import tryceratops
 from tryceratops.files import FileDiscovery, load_config
-from tryceratops.filters import GlobalSettings
 from tryceratops.interfaces import CliInterface
 from tryceratops.logging_config import LOGGING_CONFIG
 from tryceratops.runners import Runner
+from tryceratops.settings import GlobalSettings
 from tryceratops.violations import CODE_CHOICES
 
 runner = Runner()
